+++ controllers/topics-controller.js
@@ -1,4 +1,4 @@
-const { fetchTopics, fetchEndPoints, fetchArticleById } = require('../models/topics-model')
+const { fetchTopics } = require('../models/topics-model')
 
 const getTopics = ((request, response) => {
     fetchTopics().then((topic) => {
@@ -6,20 +6,4 @@ const getTopics = ((request, response) => {
       });
 })
 
-const getEndPoints = (request, response) => {
-    const endPoint = fetchEndPoints();
-    response.status(200).send({ endPoint });
-  };
-
-const getArticleById = (request, response, next) => {
-    const articleId = Number(request.params.article_id)
-    fetchArticleById(articleId).then((article) => {
-        console.log(article)
-        response.status(200).send({article})
-    })
-    .catch((err) => {
-        next(err)
-    })
-}
-
-module.exports = { getTopics, getEndPoints, getArticleById};
+module.exports = { getTopics };
