+++ models/endpoints-model.js
@@ -1,0 +1,7 @@
+const endPointData = require("../endpoints.json");
+
+const fetchEndPoints = () => {
+    return endPointData;
+  };
+
+  module.exports = {fetchEndPoints};
